requirements.txt
Authlib==1.5.2
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
cryptography==45.0.2
Flask==2.2.2
flask-cors==6.0.0
Flask-SQLAlchemy==2.5.1
google-auth==2.40.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
python-dotenv==1.1.0
requests==2.32.3
rsa==4.9.1
SQLAlchemy==1.4.46
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==2.3.8
Flask>=2.0
Flask-Cors
Flask-Login
Flask-JWT-Extended
python-dotenv
requests
gunicorn
Authlib